+++ lib/src/statements/equals.dart
@@ -1,6 +1,6 @@
 import 'package:postgres_builder/src/statements/operator_comparison.dart';
 
-class Equals extends OperatorComparision {
+class Equals extends OperatorComparison {
   const Equals(super.column, super.value)
       : super(operator: '=', columnFirst: true);
   const Equals.otherColumn(